Refactor to break import loops
* Made the data product registry load data products on request rather than at
import time to break import cycles. Now we can reference the data product
registry in various places without import loops.
* Split up app_state to allow importing subparts of the state and avoiding import loops.
